transportes_backend/requirements.txt
fastapi==0.116.1
uvicorn==0.30.1
python-dotenv==1.1.1
beanie==1.30.0
motor==3.7.1
pymongo==4.13.2
email-validator==2.2.0
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
redis==6.2.0
faker==26.0.0
pydantic-settings==2.3.3